feat: add main and api logic
